+++ src/models/EmbeddedAsset.php
@@ -382,17 +382,6 @@ class EmbeddedAsset extends Model implements JsonSerializable
     // Deprecated properties
 
     /**
-     * Method wrapper for Service::isEmbedSafe
-     *
-     * @deprecated in 2.4.0, will be removed in 3.0.0; use `getIsSafe()` instead
-     * @return bool
-     */
-    public function isSafe(): bool
-    {
-        return $this->getIsSafe();
-    }
-
-    /**
      * @return string
      * @deprecated
      */